feat(AzDO): add `fork_suffix` setting field
Azure Devops supports forks. In my $CURRENT_COMPANY, in a lot of repos,
we allow people only to create branches in their forks.

`all-repos` always tries creating branches in the same repo, as the
target branch which is problematic.

This change changes the behavior: if `fork_suffix` is defined, we are
looking for all forks of the given repository, looking for a one that
matches the `fork_suffix`.

This is because AzDO by default proposes to create a fork with
`.name.lastname` suffix when doing that throught the web interface.

If `fork_suffix` is not defined, or if no fork matching the suffix is
found, the old behavior is used (i.e. creating the branch in the same
repo, as the PR target branch).

POTENTIAL IMPROVEMENT: instead of a suffix, we can allow the user to
specify a regex against which existing forks would be checked.

Unfortunately, according to my research, it's not possible to find a
fork created/owned by the currently authorized user. That was the most
obvious choice, but this information is not available via the API.
Internet suggests to check for the initial repo commit to tell the
owner, which is a obviously wrong choice when talking about forks of
company repos.

[WIP] Version that works

fix: take care of fork_id unbound variable

refactor: try getting fork details always if fork_suffix is provided

If fork is not found, try pushing the branch directly as before.

style: pre-commit fixes

fix: test